fix(eval/padim): align dtypes and resize anomaly maps for pixel metrics
- Updated `src/eval.py`:
  - Use `torch.load(..., weights_only=False)` to properly load checkpoints with NumPy arrays in PyTorch ≥2.6.
  - Added `_align_maps_to_masks` helper to upsample anomaly maps to the ground-truth mask resolution and squeeze mask channels before metric computation.
  - Ensures `all_maps` and `all_masks` have consistent shapes for pixel-level metrics (AUROC, AUPRC, PRO).

- Updated `src/models/padim.py`:
  - Cast loaded `means` and `covs_inv` to the same `dtype` and `device` as feature embeddings inside `predict()` to fix `RuntimeError: expected scalar type Float but found Double`.

+++ src/models/padim.py
@@ -108,8 +108,8 @@ class PaDiMModel(nn.Module):
         B, C, H, W = emb.shape
         emb = emb.permute(0, 2, 3, 1).reshape(-1, C)  # [B*H*W, C]
 
-        mean = torch.from_numpy(self.means).to(self.device)
-        cov_inv = torch.from_numpy(self.covs_inv).to(self.device)
+        mean = torch.from_numpy(self.means).to(self.device).to(emb.dtype)
+        cov_inv = torch.from_numpy(self.covs_inv).to(self.device).to(emb.dtype)
 
         diffs = emb - mean
         # Mahalanobis distance squared
